Fix - Error when producing dissassembly MO
Trying to product a dissassembly MO causes the produced SQL insert
statement to fail on a database constraint related to fk_stock_movement,
the consume statement was successful. The consume statement was checking
for a '0' and setting the value to 'null' but the produce line was not
doing this check. Adding this check to fk_stock_movement fixes the error
and generates the database entries